聊天室

+++ yudao-module-bbs/yudao-module-bbs-biz/src/main/java/cn/iocoder/yudao/module/bbs/controller/app/groupmsgcontent/vo/GroupMsgContentExcelVO.java
@@ -21,7 +21,7 @@ public class GroupMsgContentExcelVO {
     private Long groupId;
 
     @ExcelProperty("发送者的编号")
-    private Integer fromId;
+    private Long fromId;
 
     @ExcelProperty("发送者的昵称")
     private String fromName;
