+++ yudao-module-bbs/yudao-module-bbs-biz/src/main/java/cn/iocoder/yudao/module/bbs/controller/app/msgcontent/vo/MsgContentExcelVO.java
@@ -21,7 +21,7 @@ public class MsgContentExcelVO {
     private String privateChatId;
 
     @ExcelProperty("发送者的编号")
-    private Integer fromId;
+    private Long fromId;
 
     @ExcelProperty("发送者的昵称")
     private String fromName;
